Add ability to update video + ProgramBlock cleanup
- Play the next video after video finishes (or go back to beginning)
- Move program block operations into its own file
