+++ src/actions/programBlockTypes.js
@@ -1,3 +1,4 @@
 export const ADD_PROGRAM_BLOCK = 'ADD_PROGRAM_BLOCK';
 export const SET_CURRENT_PROGRAM_BLOCK = 'SET_CURRENT_PROGRAM_BLOCK';
 export const UPDATE_PROGRAM_BLOCK = 'UPDATE_PROGRAM_BLOCK';
+export const SET_CURRENT_VIDEO = 'SET_CURRENT_VIDEO';
